Moved registers into gnr.web.daemon, were they should belong. Left the siteregister client inside gnrwsgisite_proxy
